move store_transaction to its own mixin, split up perf-issue tests to own class
