+++ tests/snuba/tsdb/test_tsdb_backend.py
@@ -6,6 +6,7 @@ import pytz
 from sentry.models import Environment, Group, GroupRelease, Release
 from sentry.testutils import SnubaTestCase, TestCase
 from sentry.testutils.helpers.datetime import iso_format
+from sentry.testutils.perfomance_issues.store_transaction import PerfIssueTransactionTestMixin
 from sentry.testutils.silo import region_silo_test
 from sentry.tsdb.base import TSDBModel
 from sentry.tsdb.snuba import SnubaTSDB
@@ -504,7 +505,7 @@ class SnubaTSDBTest(TestCase, SnubaTestCase):
 
 
 @region_silo_test
-class SnubaTSDBGroupPerformanceTest(TestCase, SnubaTestCase):
+class SnubaTSDBGroupPerformanceTest(TestCase, SnubaTestCase, PerfIssueTransactionTestMixin):
     def setUp(self):
         super().setUp()
